pyproject.toml
[tool.poetry]
name = "gcom"
version = "0.1.0"
description = "Ground Control Software Rewrite"
authors = ["21chanas3 <adenchan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-rest-framework = "^0.1.0"
python-socketio = "^5.11.2"
django-polymorphic = "^3.1.0"
drf-spectacular = "^0.27.2"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
